% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults}
\alias{defaults}
\alias{defaultCmakeDownloadVersion}
\alias{defaultCmakeMinimumVersion}
\alias{defaultCmakeCacheDirectory}
\title{Defaults for \pkg{biocmake}}
\usage{
defaultCmakeDownloadVersion()

defaultCmakeMinimumVersion()

defaultCmakeCacheDirectory()
}
\value{
For \code{defaultCmakeCommand}, a string specifying the expected command-line invocation of an existing Cmake installation.

For \code{defaultCmakeDownloadVersion}, a string specifying the version of Cmake to download if no existing installation can be found.

For \code{defaultCmakeMinimumVersion}, a string specifying the minimum version of an existing Cmake installation.

For \code{defaultCmakeCacheDirectory}, a string containing the path to the cache directory for \pkg{biocmake}-managed Cmake installations.
}
\description{
Defaults for \pkg{biocmake}
}
\details{
The \code{BIOCMAKE_CMAKE_COMMAND} environment variable will override the default setting of \code{defaultCmakeCommand}.

The \code{BIOCMAKE_CMAKE_DOWNLOAD_VERSION} environment variable will override the default setting of \code{defaultCmakeDownloadVersion}.

The \code{BIOCMAKE_CMAKE_MINIMUM_VERSION} environment variable will override the default setting of \code{defaultCmakeMinimumVersion}.

The \code{BIOCMAKE_CMAKE_CACHE_DIRECTORY} environment variable will override the default setting of \code{defaultCmakeCacheDirectory}.
}
\examples{
defaultCmakeCommand()
defaultCmakeDownloadVersion()
defaultCmakeMinimumVersion()
defaultCmakeCacheDirectory()

}
\author{
Aaron Lun
}
