% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Download Cmake}
\usage{
download(
  download.version = defaultDownloadVersion(),
  cache.dir = defaultCacheDirectory(),
  ignore.cache = FALSE
)
}
\arguments{
\item{download.version}{String specifying the Cmake version to download.}

\item{cache.dir}{String specifying the location of the directory in which to cache Cmake installations.}

\item{ignore.cache}{Logical scalar specifying whether to ignore any existing cached version of Cmake,
in which case the binaries will be downloaded again.}
}
\value{
String containing the path to the Cmake executable.
}
\description{
Download Cmake binaries for the current architecture.
}
\examples{
download()

}
\author{
Aaron Lun
}
