% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCmake.R
\name{findCmake}
\alias{findCmake}
\title{Find Cmake}
\usage{
findCmake(
  command = defaultCmakeCommand(),
  minimum.version = defaultCmakeMinimumVersion(),
  download = TRUE,
  ...
)
}
\arguments{
\item{command}{String containing the command to check for an existing installation.}

\item{minimum.version}{String specifying the minimum acceptable version of an existing installation.}

\item{download}{Logical scalar indicating whether to download Cmake if no acceptable existing installation can be found.}

\item{...}{Further arguments to pass to \code{\link{downloadCmake}}.}
}
\value{
String containing the command to use to run Cmake.
}
\description{
Find an existing Cmake installation or, if none can be found, install a \pkg{biocmake}-managed Cmake instance.
}
\examples{
cmd <- findCmake()
system2(cmd, "--version")

}
\author{
Aaron Lun
}
