% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{configure}
\alias{configure}
\title{Configure Cmake}
\usage{
configure(compact = TRUE)
}
\arguments{
\item{compact}{Logical scalar indicating whether to format the options as Cmake command-line arguments.}
}
\value{
If \code{compact}, a character vector containing CMake command-line arguments.

Otherwise, a named character vector containing the name and value of each option.
}
\description{
Propagate R's configuration variables into the Cmake options, where possible.
}
\examples{
configure()

}
\author{
Aaron Lun
}
