% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults}
\alias{defaults}
\alias{defaultDownloadVersion}
\alias{defaultMinimumVersion}
\alias{defaultCacheDirectory}
\title{Defaults for \pkg{biocmake}}
\usage{
defaultDownloadVersion()

defaultMinimumVersion()

defaultCacheDirectory()
}
\value{
For \code{defaultCommand}, a string specifying the expected command-line invocation of an existing Cmake installation.

For \code{defaultDownloadVersion}, a string specifying the version of Cmake to download if no existing installation can be found.

For \code{defaultMinimumVersion}, a string specifying the minimum version of an existing Cmake installation.

For \code{defaultCacheDirectory}, a string containing the path to the cache directory for \pkg{biocmake}-managed Cmake installations.
}
\description{
Defaults for \pkg{biocmake}
}
\details{
The \code{BIOCMAKE_CMAKE_COMMAND} environment variable will override the default setting of \code{defaultCommand}.

The \code{BIOCMAKE_CMAKE_DOWNLOAD_VERSION} environment variable will override the default setting of \code{defaultDownloadVersion}.

The \code{BIOCMAKE_CMAKE_MINIMUM_VERSION} environment variable will override the default setting of \code{defaultMinimumVersion}.

The \code{BIOCMAKE_CMAKE_CACHE_DIRECTORY} environment variable will override the default setting of \code{defaultCacheDirectory}.
}
\examples{
defaultCommand()
defaultDownloadVersion()
defaultMinimumVersion()
defaultCacheDirectory()

}
\author{
Aaron Lun
}
